more error handling2

+++ models/comment.js
@@ -1,30 +1,34 @@
 const connection = require("../db/connection");
-const { fetchArticleById } = require("./article");
+const { checkArticle } = require("./article");
 
 exports.addComment = (comment) => {
   const { article_id } = comment;
 
-  console.log(article_id);
-
-  return connection("comments")
-    .insert(comment)
-    .returning("*")
-    .then(([response]) => {
-      return response;
-    });
+  return checkArticle(article_id).then((response) => {
+    if (!response) Promise.reject();
+    return connection("comments")
+      .insert(comment)
+      .returning("*")
+      .then(([response]) => {
+        return response;
+      });
+  });
 };
 
 exports.fetchCommentsById = (
   { article_id },
   { sort_by = "created_at", order_by = "asc" }
 ) => {
-  return connection("comments")
-    .select("*")
-    .where("article_id", article_id)
-    .orderBy(sort_by, order_by)
-    .then((comments) => {
-      return comments;
-    });
+  return checkArticle(article_id).then((response) => {
+    if (!response) Promise.reject();
+    return connection("comments")
+      .select("*")
+      .where("article_id", article_id)
+      .orderBy(sort_by, order_by)
+      .then((comments) => {
+        return comments;
+      });
+  });
 };
 
 exports.modifyComment = ({ inc_votes }, { comment_id }) => {
@@ -45,6 +49,24 @@ exports.modifyComment = ({ inc_votes }, { comment_id }) => {
     });
 };
 
+const checkComment = (comment_id) => {
+  if (isNaN(comment_id)) return false;
+  return connection("comments")
+    .select("*")
+    .where("comment_id", comment_id)
+    .then((comments) => {
+      if (Object.keys(comments[0]).length === 0) return false;
+      return true;
+    });
+};
+
 exports.removeComment = ({ comment_id }) => {
-  return connection("comments").where("comment_id", comment_id).delete();
+  return checkComment(comment_id).then((response) => {
+    if (!response) {
+      console.log("inside");
+      Promise.reject();
+    } else {
+      return connection("comments").where("comment_id", comment_id).delete();
+    }
+  });
 };
